Removed old enum and functions/added dynamic array
Removed old enum from MainActivity (not being used), and removed the
getToday function from the enum class it was copied to.  Also created a
public global array to hold our 7 day week, which is determined by the
current day of the week.  Created a function to build this array
dynamically based on what day it is.  Will be used for interface
ordering.  Still need to do code cleanup/warning elimination.