New FileManager and JSON done
Finalized the retrieval/formatting of remote JSON data and created a
new FileManager class to handle saving and retrieving this data to
local storage.  Working on creating the necessary checks within
MainActivity to handle old data/no data and recreating the interface
around the new JSON that is pulled in.