Interface Loading dynamic data
Interface now loads data that is either pulled from online, or loaded
from device storage.  All interface elements are utilizing data
appropriately.  Still need to work on checks required for current data,
and if a file already exists.  Also need to create another array to
hold an unformatted version of data for the primary “weather_holder”
view, as currently it shares the shortened string with the grid view.