Added dynamic images
Loaded all images from API and converted to byteArrays, and then used
the Base64 class to encode them to strings for storage within JSON.
This way, the user isn’t constantly redownloading the images if they
don’t have to.  Still need to correctly position the imageView, but
since the interface is going to mostly be redone this week, figured it
best to wait for now.