Fixed bug where emulator may crash due to GPS
Fixed a bug where the app may crash due to the fact that the emulator
didn’t play nice with the GPS functionality.  We now set the location
manually for an instances where the app is run within the emulator.
Cannot recreate the crashing reported by having wifi disabled while
tapping “Refresh”, even when toggling network within the emulator.